[Refactor] Move DeepSeek V4 attention.py out of nvidia/ subdir
The DeepseekV4 MLA attention layer already handles both NVIDIA and ROCm
via current_platform.is_rocm() dispatch (and a TYPE_CHECKING-only import
from nvidia.flashmla); its home under nvidia/ops/ was a historical
artifact. Move it to the deepseek_v4/ package root alongside the other
hardware-agnostic modules (compressor.py, quant_config.py).

Update the four importers (nvidia/model.py, nvidia/flashmla.py,
amd/rocm.py, and tests/kernels/test_fused_inv_rope_fp8_quant.py docstrings).
amd/model.py is a symlink to nvidia/model.py so it inherits the change.

Verified by running tests/kernels/test_fused_deepseek_v4_qnorm_rope_kv_insert.py
(139 passed).
